pyproject.toml
[tool.poetry]
name = "api-menu"
version = "0.1.0"
description = ""
authors = ["andSpt <henrrry52@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "==0.108.0"
uvicorn = {extras = ["standart"], version = "==0.25.0"}
sqlalchemy = {extras = ["asyncio"], version = "==2.0.25"}
asyncpg = "==0.29.0"
pydantic-settings = "==2.1.0"
httpx = "==0.26.0"
black = "==24.2.0"
alembic = "==1.13.1"
celery = "==5.3.6"
rabbitmq = "==0.2.0"
email-validator = "==2.1.1"
pyjwt = {extras = ["crypto"], version = "==2.8.0"}
bcrypt = "==4.1.2"
python-multipart = "==0.0.9"
sqlmodel = "==0.0.16"
pytest-asyncio = "0.18.3"
pytest-order = "==1.2.0"
redis = "==5.2.0"
isort = "^6.0.1"

[tool.pytest.ini_options]
testpaths = "app/tests"
addopts = '''
  --asyncio-mode=auto
'''

[tool.poetry.group.dev.dependencies]
pytest = "==7.4.4"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
